Added Temporary Layouts for testing.
Temporary layouts will load a keyboard layout for testing purposes.
Once the test is complete, if the layout is no longer needed, the
layout will be unloaded. The user's layout will be returned once the
temporary layout is disposed of.

A Windows implementation has been added with this initial prototype.
This feature has been implemented in a way that keeps the door open for
cross-platform compatibility in the future.

I also deleted a bunch of stuff in the `User32` imports class, as it
was not being used.