Implemented runtime loading.
I've implemented runtime loading. The Logitech DLLs no longer need to
be in the KSP root folder, instead, they only need to be in the
relative folder Logitech/x64 or Logitech/x86. The code automatically
determines which version to use. Also renamed the `ScanCode` enum to
`KeyName` after talking with the Logitech SDK team.