Licensing my individual contributions under APL-2.0
I've licensed my individual non-KSP-specific contributions under the
Apache License. This is for a few specific reasons:

1. I would like my code to be usable by indie developers looking to
   implement RGB lighting in their own games.
2. I am considering spinning off the keyboard layout handling code into
   a separate library licensed under APL-2.0.
3. The mod itself will still be licensed under GPLv3, any KSP-specific
   code should and will remain licensed under the GPLv3. This will
   protect against Take-Two using the code without permission, or those
   attempting to sell a pre-packaged version of the mod.

I believe this strikes a fair balance between protecting the work of
contributors and opening up the source to indie developers. The "crown
jewels" of this plugin is and always will be the specific KSP
integrations, which will remain to be protected under the stricter
terms of the GPLv3 license.