Removed log messages from KSP not recognising native DLLs.
KSP was having issues when trying to load the Logitech DLLs, causing a
spurious log message in KSP.log. I have fixed the issue by renaming
the DLLs to **LogitechLedEnginesWrapper.nativedll.(x64|x86)**, and
moving them directly to the Logitech directory. By putting the processor
architecture in the filename, I have eliminated the need for separate
folders for each processor platform.

The Logitech folder will need to be copied to the GameData/KSPLogiRGB
folder and the test binaries directory.